Bugfix: BEEEEEEEEES!
Because of the previous fix for missing bees there were now more bees.
Many more. As in. A lot. Really! Soooo many! BEES!

Happened because bees were added to all comment section instead of the
one which reloaded.

Adding a preselector to the addQuickFeedback method allows for targeting
a smaller subset of HTML nodes. Now the post id of the comment reload
url is used as id preselector.