Request to Higgs for more precise feedback
The feedback options nc, fp and sk are only relevant if a post is
reported. Now there is a check before sending those feedbacks which
makes sure that a comment is reported.

A tp will always be sent no matter whether it was reported or not as
those are always relevant.