YAAAAY! NAAAAY!
When flagging a comment as "No longer needed" a small dialog will pop up
asking whether that comment is considered not constructive with the
options YAY! and NAY!. Yay sends "nc" as feedback and Nay sends "fp" as
feedback to Queen. Clicking outside of the dialog or the X in the corner
closes it like a normal SO popup. This popup cannot be moved and not be
closed by hitting Esc. Sorry ^^"